requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
opencv-python==4.8.0.76
pillow==10.0.0
pytesseract==0.3.10
pywin32==306
python-dateutil==2.8.2 